+++ app/Http/Requests/PatchUserRequest.php
@@ -7,7 +7,7 @@ use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Lang;
 
-class EditUserRequest extends FormRequest
+class PatchUserRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -39,6 +39,6 @@ class EditUserRequest extends FormRequest
      */
     public function messages()
     {
-        return Lang::get('validation/editUser');
+        return Lang::get('validation/patchUser');
     }
 }