+++ changelog.md
@@ -23,6 +23,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Add inventory item request.
     * Patch inventory item request.
     * Delete inventory item request.
+    
+### Changed
+
+* Edit user request
+    * Renamed to Patch user request in controller, tests and language files.
 
 ### Removed